move esm-2 to flattened requirements.txt layout

remove session-scoped checkpoint fixture


address coderabbit review
