+++ bionemo-recipes/models/esm2/.ruff.toml
@@ -1,0 +1,1 @@
+extend = "../.ruff.toml"
