+++ bionemo-recipes/models/esm2/Dockerfile
@@ -2,4 +2,4 @@ FROM nvcr.io/nvidia/pytorch:26.01-py3
 WORKDIR /workspace/bionemo
 COPY . .
 RUN --mount=type=cache,target=/root/.cache/pip \
-    PIP_CONSTRAINT= pip install -e .
+    PIP_CONSTRAINT= pip install -r requirements.txt
