+++ bionemo-recipes/models/esm2/collator.py
