+++ bionemo-recipes/models/esm2/tests/conftest.py
@@ -25,11 +25,10 @@ from transformer_engine.common import recipe as recipe_module
 from transformer_engine.pytorch import fp8
 from transformers import AutoModelForMaskedLM, AutoTokenizer, DataCollatorForLanguageModeling
 
-from esm.convert import convert_esm_hf_to_te
-
 
 sys.path.append(Path(__file__).parent.parent.as_posix())
 sys.path.append(Path(__file__).parent.as_posix())
+
 
 pytest_plugins = ["tests.common.fixtures"]
 
@@ -105,6 +104,8 @@ def input_data(tokenizer, tokenized_proteins):
 
 @pytest.fixture
 def te_model_checkpoint(tmp_path):
+    from convert import convert_esm_hf_to_te
+
     model_hf = AutoModelForMaskedLM.from_pretrained("facebook/esm2_t6_8M_UR50D", revision="c731040f")
     model_te = convert_esm_hf_to_te(model_hf)
     model_te.save_pretrained(tmp_path / "te_model_checkpoint")
