+++ bionemo-recipes/models/esm2/tests/test_cp_bshd.py
@@ -15,18 +15,24 @@
 
 import os
 import subprocess
+import sys
 import tempfile
 from dataclasses import dataclass, field
 from pathlib import Path
+
+
+# When launched via torchrun, conftest.py sys.path setup doesn't run.
+# Ensure the model directory (parent of tests/) is on sys.path for bare module imports.
+sys.path.insert(0, Path(__file__).resolve().parent.parent.as_posix())
 
 import pytest
 import torch
 from torch.distributed.device_mesh import init_device_mesh
 from transformers import AutoModelForMaskedLM, AutoTokenizer, DataCollatorForLanguageModeling
 
-from esm.collator import _split_batch_by_cp_rank
-from esm.convert import convert_esm_hf_to_te
-from esm.modeling_esm_te import NVEsmForMaskedLM
+from collator import _split_batch_by_cp_rank
+from convert import convert_esm_hf_to_te
+from modeling_esm_te import NVEsmForMaskedLM
 
 
 requires_multi_gpu = pytest.mark.skipif(
